// src/types/supabase.ts
export interface MyProject {
    id: number;
    title: string;
    reason_created: string;
    trouble_shooting: string;
    features: string;
    part: number;
    github: string;
    url: string;
    technologies: string;
    start_date: string;
    end_date: string;
    more: string;
    type: string;
}